chore: remove dependency from Store in backfill_sync
